// CSKT/originaldb-gate-fe/src/modules/it-categories/types/EngineRoom.type.ts
export type EngineRoom = {
  id: string;
  name: string;
  short_name: string;
  phone_number: string;
  contact_method: string;
  org_id: string;
  org_use_id: string;
  org_technical_id: string;
  org_support_id: string;
  ward_id: any;
  address: string;
  longitude: string;
  latitude: string;
  area: number;
  n_phy_server: number;
  n_vir_server: number;
  sum_app_service: number;
  ability_cpu: number;
  sum_ram: number;
  bandwidth: number;
  n_storage: number;
  n_rack: number;
  n_percent_density: number;
  comment: string;
  url_file_path: string;
  order_number: number;
  has_fighting_protect: boolean;
  has_lighting_protect: boolean;
  has_power_source: boolean;
  is_enable: boolean;
};